+++ server/impl/src/main/java/com/walmartlabs/concord/server/security/sso/SsoToken.java
@@ -1,3 +1,5 @@
+package com.walmartlabs.concord.server.security.sso;
+
 /*-
  * *****
  * Concord
@@ -7,9 +9,9 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * 
  *      http://www.apache.org/licenses/LICENSE-2.0
- *
+ * 
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -17,10 +19,35 @@
  * limitations under the License.
  * =====
  */
-import { deepMerge } from '../common';
 
-test('deepMerge works', () => {
-    const actual = deepMerge({ x: { y: 123 } }, { x: { z: 234 } });
-    const expected = { x: { y: 123, z: 234 } };
-    expect(actual).toEqual(expected);
-});
+import org.apache.shiro.authc.AuthenticationToken;
+
+public class SsoToken implements AuthenticationToken {
+
+    private final String login;
+
+    public SsoToken(String login) {
+        this.login = login;
+    }
+
+    public String getLogin() {
+        return login;
+    }
+
+    @Override
+    public Object getPrincipal() {
+        return login;
+    }
+
+    @Override
+    public Object getCredentials() {
+        return login;
+    }
+
+    @Override
+    public String toString() {
+        return "SsoToken{" +
+                "login='" + login + '\'' +
+                '}';
+    }
+}